pyproject.toml
[project]
name = "kata"
version = "0.1.0"
description = "Terminal-centric workspace orchestrator for tmux"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Kata Contributors" }]
keywords = ["tmux", "terminal", "workspace", "orchestrator", "cli", "tui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Terminals :: Terminal Emulators/X Terminals",
]

dependencies = [
    "textual>=0.47.0",
    "typer>=0.9.0",
    "tmuxp>=1.47.0",
    "libtmux>=0.37.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
kata = "kata.cli.app:app"

[project.urls]
Homepage = "https://github.com/kata-workspace/kata"
Documentation = "https://github.com/kata-workspace/kata#readme"
Repository = "https://github.com/kata-workspace/kata"
Issues = "https://github.com/kata-workspace/kata/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kata"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["kata"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
strict = true

[[tool.mypy.overrides]]
module = ["tmuxp.*", "libtmux.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["kata"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
